fix(230): verify menu state helpers

+++ game/230/230_42_MM_Characters_boolIsInvalid.c
@@ -2,9 +2,10 @@
 
 // used for preventing players highlighting the same character
 // also for when you go left of komodo joe's icon
+// NOTE(aalhendi): ASM-verified against NTSC-U 926 overlay 230 0x800adae4-0x800adb64.
 u32 MM_Characters_boolIsInvalid(s16 *globalIconPerPlayer, s16 characterID, s16 player)
 {
-	int i = 0;
+	s16 i = 0;
 
 	// if there are players
 	if (sdata->gGT->numPlyrNextGame)
